refactor move tabState into store/ directory
