+++ src/pageEditor/TabConnectionErrorBanner.tsx
@@ -19,7 +19,7 @@ import React from "react";
 import { Button } from "react-bootstrap";
 import { getErrorMessage, isSpecificError } from "@/errors/errorHelpers";
 import { useSelector } from "react-redux";
-import { selectTabStateError } from "@/pageEditor/tabState/tabStateSelectors";
+import { selectTabStateError } from "@/pageEditor/store/tabState/tabStateSelectors";
 import { TimeoutError } from "p-timeout";
 
 /**
