+++ src/pageEditor/modListingPanel/AddStarterBrickButton.tsx
@@ -23,7 +23,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { sortBy } from "lodash";
 import useAddNewModComponent from "@/pageEditor/hooks/useAddNewModComponent";
 import { useSelector } from "react-redux";
-import { selectTabHasPermissions } from "@/pageEditor/tabState/tabStateSelectors";
+import { selectTabHasPermissions } from "@/pageEditor/store/tabState/tabStateSelectors";
 import useAsyncState from "@/hooks/useAsyncState";
 import { faExternalLinkAlt } from "@fortawesome/free-solid-svg-icons";
 import reportEvent from "@/telemetry/reportEvent";
