+++ src/pageEditor/store/store.ts
@@ -37,7 +37,7 @@ import { logSlice } from "@/components/logViewer/logSlice";
 import { authSlice, persistAuthConfig } from "@/auth/authSlice";
 import analysisSlice from "@/analysis/analysisSlice";
 import pageEditorAnalysisManager from "../analysisManager";
-import { tabStateSlice } from "@/pageEditor/tabState/tabStateSlice";
+import { tabStateSlice } from "@/pageEditor/store/tabState/tabStateSlice";
 import { modDefinitionsSlice } from "@/modDefinitions/modDefinitionsSlice";
 import { modDefinitionsMiddleware } from "@/modDefinitions/modDefinitionsListenerMiddleware";
 import {
