+++ src/pageEditor/store/tabState/tabStateSelectors.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { type TabStateRootState } from "@/pageEditor/tabState/tabStateTypes";
+import { type TabStateRootState } from "@/pageEditor/store/tabState/tabStateTypes";
 import { deserializeError } from "serialize-error";
 
 export const selectTabStateError = ({ tabState }: TabStateRootState) =>
