+++ src/tsconfig.strictNullChecks.json
@@ -1157,9 +1157,9 @@
     "./pageEditor/starterBricks/quickBarProvider.ts",
     "./pageEditor/starterBricks/sidebar.ts",
     "./pageEditor/starterBricks/trigger.ts",
-    "./pageEditor/tabState/tabStateSelectors.ts",
-    "./pageEditor/tabState/tabStateSlice.ts",
-    "./pageEditor/tabState/tabStateTypes.ts",
+    "./pageEditor/store/tabState/tabStateSelectors.ts",
+    "./pageEditor/store/tabState/tabStateSlice.ts",
+    "./pageEditor/store/tabState/tabStateTypes.ts",
     "./pageEditor/tabs/ExtraPermissionsSection.tsx",
     "./pageEditor/tabs/Logs.tsx",
     "./pageEditor/tabs/MatchRulesSection.tsx",
